+++ forui/example/lib/sandbox.dart
@@ -64,7 +64,7 @@ class _SandboxState extends State<Sandbox> {
             label: const Text('Select Group'),
             description: const Text('Select Group Description'),
             controller: FMultiSelectGroupController(min: 1, max: 2, values: {1}),
-            items: [
+            children: [
               FSelectGroupItem.checkbox(value: 1, label: const Text('Checkbox 1'), semanticLabel: 'Checkbox 1'),
               FSelectGroupItem.checkbox(value: 2, label: const Text('Checkbox 2'), semanticLabel: 'Checkbox 2'),
               FSelectGroupItem.checkbox(value: 3, label: const Text('Checkbox 3'), semanticLabel: 'Checkbox 3'),