+++ forui/test/src/widgets/select_group/select_group_golden_test.dart
@@ -9,7 +9,42 @@ import 'package:forui/forui.dart';
 import '../../test_scaffold.dart';
 
 void main() {
-  group('FCard', () {
+  group('FCheckbox', () {
+    testWidgets('blue screen', (tester) async {
+      await tester.pumpWidget(
+        TestScaffold.blue(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              SizedBox(
+                width: 300,
+                child: FSelectGroup(
+                  style: TestScaffold.blueScreen.selectGroupStyle,
+                  label: const Text('Select Group'),
+                  description: const Text('Select Group Description'),
+                  controller: FMultiSelectGroupController(values: {1}),
+                  children: [
+                    FSelectGroupItem.checkbox(
+                      value: 1,
+                      label: const Text('Checkbox 1'),
+                      semanticLabel: 'Checkbox 1',
+                    ),
+                    FSelectGroupItem.radio(
+                      value: 2,
+                      label: const Text('Checkbox 2'),
+                      semanticLabel: 'Checkbox 2',
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ),
+      );
+
+      await expectLater(find.byType(TestScaffold), matchesGoldenFile('blue-screen.png'));
+    });
+    
     for (final (name, theme, background) in TestScaffold.themes) {
       testWidgets('$name with checkbox', (tester) async {
         await tester.pumpWidget(
@@ -25,7 +60,7 @@ void main() {
                     label: const Text('Select Group'),
                     description: const Text('Select Group Description'),
                     controller: FMultiSelectGroupController(values: {1}),
-                    items: [
+                    children: [
                       FSelectGroupItem.checkbox(
                         value: 1,
                         label: const Text('Checkbox 1'),
@@ -49,10 +84,7 @@ void main() {
           ),
         );
 
-        await expectLater(
-          find.byType(TestScaffold),
-          matchesGoldenFile('select-group/$name-checkbox.png'),
-        );
+        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name-checkbox.png'));
       });
 
       testWidgets('$name with checkbox error', (tester) async {
@@ -70,7 +102,7 @@ void main() {
                     description: const Text('Select Group Description'),
                     forceErrorText: 'Some error message.',
                     controller: FMultiSelectGroupController(values: {1}),
-                    items: [
+                    children: [
                       FSelectGroupItem.checkbox(
                         value: 1,
                         label: const Text('Checkbox 1'),
@@ -94,10 +126,7 @@ void main() {
           ),
         );
 
-        await expectLater(
-          find.byType(TestScaffold),
-          matchesGoldenFile('select-group/$name-checkbox-error.png'),
-        );
+        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name-checkbox-error.png'));
       });
     }
   });
@@ -118,7 +147,7 @@ void main() {
                     label: const Text('Select Group'),
                     description: const Text('Select Group Description'),
                     controller: FRadioSelectGroupController(value: 1),
-                    items: [
+                    children: [
                       FSelectGroupItem.radio(
                         value: 1,
                         label: const Text('Radio 1'),
@@ -142,10 +171,7 @@ void main() {
           ),
         );
 
-        await expectLater(
-          find.byType(TestScaffold),
-          matchesGoldenFile('select-group/$name-radio.png'),
-        );
+        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name-radio.png'));
       });
 
       testWidgets('$name with radio error', (tester) async {
@@ -163,7 +189,7 @@ void main() {
                     description: const Text('Select Group Description'),
                     forceErrorText: 'Some error message.',
                     controller: FRadioSelectGroupController(value: 1),
-                    items: [
+                    children: [
                       FSelectGroupItem.radio(
                         value: 1,
                         label: const Text('Radio 1'),
@@ -187,10 +213,7 @@ void main() {
           ),
         );
 
-        await expectLater(
-          find.byType(TestScaffold),
-          matchesGoldenFile('select-group/$name-radio-error.png'),
-        );
+        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name-radio-error.png'));
       });
     }
   });